Move extensibility tests to separate file.
